# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HealthManager API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/patients": {
      "get": {
        "operationId": "patients_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortDirection",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "healthInsurance",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedPatientResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "patients_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}": {
      "patch": {
        "operationId": "patients_update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "patients_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/patients/{id}/access-token/regenerate": {
      "post": {
        "operationId": "patientsRegenerateAccessToken",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/documents": {
      "get": {
        "operationId": "patients_documents_list",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDocumentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "patients_documents_create",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDocumentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/documents/upload": {
      "post": {
        "operationId": "patients_documents_upload",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadPatientDocumentForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDocumentResponse"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/documents/{documentId}/download": {
      "get": {
        "operationId": "patients_documents_download",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{id}/documents/{documentId}": {
      "delete": {
        "operationId": "patients_documents_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "documentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/appointments": {
      "get": {
        "operationId": "appointments_list",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "doctorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "Scheduled",
                "Confirmed",
                "Cancelled",
                "Completed",
                "NoShow",
                "InProgress"
              ]
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedAppointmentResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "appointments_create",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAppointmentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/portal/auth": {
      "post": {
        "operationId": "portal_auth",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientPortalLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPortalAuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portal/me": {
      "get": {
        "operationId": "portal_me",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPortalProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/portal/appointments": {
      "get": {
        "operationId": "portal_appointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPortalAppointmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/portal/receivables": {
      "get": {
        "operationId": "portal_receivables",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientPortalReceivableResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/portal/documents": {
      "get": {
        "operationId": "portal_documents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientDocumentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/patients/{patientId}/clinical-records": {
      "get": {
        "operationId": "clinical_record_list_by_patient",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClinicalRecordResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePatientRequest": {
        "type": "object",
        "required": [
          "name",
          "cpf",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "healthInsurance": {
            "type": "string"
          },
          "healthInsuranceId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/PatientDetails"
          }
        }
      },
      "CreatePatientDocumentRequest": {
        "type": "object",
        "required": [
          "fileName",
          "contentType",
          "sizeInBytes"
        ],
        "properties": {
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeInBytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "UploadPatientDocumentForm": {
        "type": "object",
        "required": [
          "file"
        ],
        "properties": {
          "file": {
            "type": "string",
            "format": "binary"
          }
        }
      },
      "PatientDocumentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeInBytes": {
            "type": "integer",
            "format": "int64"
          },
          "storagePath": {
            "type": "string"
          }
        }
      },
      "PatientResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "healthInsurance": {
            "type": "string"
          },
          "healthInsuranceId": {
            "type": "string",
            "format": "uuid"
          },
          "healthInsuranceName": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "patientAccessToken": {
            "type": "string",
            "format": "uuid"
          },
          "details": {
            "$ref": "#/components/schemas/PatientDetails"
          }
        }
      },
      "UpdatePatientRequest": {
        "type": "object",
        "required": [
          "name",
          "phone"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "healthInsurance": {
            "type": "string"
          },
          "healthInsuranceId": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "details": {
            "$ref": "#/components/schemas/PatientDetails"
          }
        }
      },
      "PatientDetails": {
        "type": "object",
        "properties": {
          "socialName": {
            "type": "string"
          },
          "rg": {
            "type": "string"
          },
          "sex": {
            "type": "string",
            "enum": [
              "Masculino",
              "Feminino",
              "Prefiro não informar"
            ]
          },
          "secondaryPhone": {
            "type": "string"
          },
          "commercialPhone": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "medicalRecordNumber": {
            "type": "string"
          },
          "healthInsuranceNumber": {
            "type": "string"
          },
          "cns": {
            "type": "string"
          },
          "isVip": {
            "type": "boolean"
          },
          "excludeFromMarketing": {
            "type": "boolean"
          },
          "receiveDirectMail": {
            "type": "boolean"
          },
          "tags": {
            "type": "string"
          },
          "zipCode": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "complement": {
            "type": "string"
          },
          "neighborhood": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "acquisitionSource": {
            "type": "string"
          },
          "referredBy": {
            "type": "string"
          },
          "maritalStatus": {
            "type": "string"
          },
          "education": {
            "type": "string"
          },
          "profession": {
            "type": "string"
          },
          "religion": {
            "type": "string"
          },
          "fatherName": {
            "type": "string"
          },
          "motherName": {
            "type": "string"
          },
          "companionName": {
            "type": "string"
          },
          "childrenCount": {
            "type": "integer"
          },
          "spouseName": {
            "type": "string"
          }
        }
      },
...
      "PagedPatientResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "PagedAppointmentResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AppointmentResponse"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
...
      "CreateAppointmentRequest": {
        "type": "object",
        "required": [
          "patientId",
          "doctorId",
          "startAt",
          "appointmentTypeId"
        ],
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "durationMinutes": {
            "type": "integer",
            "default": 30
          },
          "notes": {
            "type": "string"
          },
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "AppointmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "confirmationStatus": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "appointmentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "decimal"
          },
          "notes": {
            "type": "string"
          },
          "patientName": {
            "type": "string"
          },
          "patientPhone": {
            "type": "string"
          },
          "doctorName": {
            "type": "string"
          },
          "doctorSpecialty": {
            "type": "string"
          }
        }
      },
...
      "PatientPortalLoginRequest": {
        "type": "object",
        "required": [
          "cpf",
          "accessToken"
        ],
        "properties": {
          "cpf": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        }
      },
      "PatientPortalProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "cpf": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "healthInsurance": {
            "type": "string"
          }
        }
      },
      "PatientPortalAuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientPortalProfileResponse"
          }
        }
      },
      "PatientPortalAppointmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "doctorName": {
            "type": "string"
          },
          "doctorSpecialty": {
            "type": "string"
          },
          "startAt": {
            "type": "string",
            "format": "date-time"
          },
          "endAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          }
        }
      },
      "PatientPortalReceivableResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "originalAmount": {
            "type": "number"
          },
          "receivedAmount": {
            "type": "number"
          },
          "outstandingAmount": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ClinicalRecordResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "appointmentId": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "doctorId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "Draft",
              "Finalized"
            ]
          },
          "chiefComplaint": {
            "type": "string",
            "nullable": true
          },
          "history": {
            "type": "string",
            "nullable": true
          },
          "physicalExam": {
            "type": "string",
            "nullable": true
          },
          "assessment": {
            "type": "string",
            "nullable": true
          },
          "plan": {
            "type": "string",
            "nullable": true
          },
          "finalizedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "patientName": {
            "type": "string",
            "nullable": true
          },
          "doctorName": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  }
}